Docs/fileio (#30)
* Converting python file operations to Bash

As the program needs to be operated without user interaction, a sudoers file needs to be used to allow file operations to occur. The sudoers file only allows for programs. When DisTRaX is installed the file will be installed in the users local space thus it is not safe to allow this to be added to the Sudoers file as  it could be modified. Therefore, the file operations have been modified to bash as these can be added to the Sudoers file.

* Updating Tests

Ensuring the correct mode is reported, and adding the admin option to the tests requiring sudo.