Docs/mpi (#59)
* Creation MPI

This is a draft it needs a clean.

* check_host -> hosts_are_different

Made the function clearer and easier to understand.

* mpi_removal function

This allows for a dictionary to be passed with all the components to remove, it then removes them in the desired order with mpi.

* Adding MPI version of DisTRaX

* Fix/networking (#52)

* The tests did not properly capture the correct information.

So for the ip address 127.0.0.1 the netmask should be 8 instead of 32.

* Updated function to include ip_address nd the netmask did not properly capture the correct information.

This passes the updated tests.

* Updated docs to reflect changes.

* Updated ceph-mon to reflect network changes.

* Fix/osds spin lock (#53)

* Updating abstract_class

* Added a spin lock to ensure the OSDs are remove properly.

This checks that the correct return code is returned before moving onto the next process for removeal. This was added as when in MPI sometimes the OSDs did not remove properly.

* Fix/change ownership fileio (#54)

* Updated distrax-ceph

Python cannot see folders that are protected or owned by sudo unless running as sudo, therefore the bash command is needed.

* Fixed recursive_change_ownership

Python cannot see folders that are protected or owned by sudo unless running as sudo, therefore the bash command is needed.

* If the output is zero return a dict full of zeros. (#56)

* Interface Does Not Exist Error (#57)

* Fix/filesystem (#58)

* Change to getpass

getpass is used os.getlogin() requires a shell to operate which is not the case in a slurm submitted job whereas getpass.getuser() does not.

* Ensure the mount exists

Previously this assumed the mount would be there but in cases where the system fails the mount will not be present.

* Rank 0 Services

Previously the services where deployed on every node which does not work effectively. Therefore, they are now deployed on the Rank 0 Host.

* Additional Logging

* Additional Logging

* Adding time out

Incase a user cannot be created a timeout error has been included.

* Removed unneeded print statement.

* Documentation